# REFACTOR - MessageHandler -> MessageBuilder - Generally speaking, `MessageHandler` should work as listening to an event and handle it (think about nodejs event processing, GUI event handler such as MFC, Java Swing)   + So rename the class   + For polymorphism, only `build` function is allowed. The caller doesn't need to know numbers of functions corresponding to the message type - `channel_interface/InNetMsg, OutNetMsg` moves to `message.hpp`   + Because those are related to `message,` not `channel_interface.`
