+++ src/plugins/channel_interface/include/channel_interface.hpp
@@ -8,19 +8,6 @@
 using namespace gruut::net_plugin;
 using namespace std;
 
-// TODO: need to separate data structure and channel.
-struct InNetMsg {
-  MessageType type;
-  nlohmann::json body;
-  sender_id_type sender_id;
-};
-
-struct OutNetMsg {
-  MessageType type;
-  nlohmann::json body;
-  vector<sender_id_type> receivers;
-};
-
 namespace appbase::incoming {
 namespace channels {
 using network = ChannelTypeTemplate<InNetMsg>;
